+++ src/client/components/ThemaIcon/ThemaIcon.test.tsx
@@ -3,7 +3,7 @@ import ThemaIcon from './ThemaIcon';
 import { Themas } from '../../../universal/config';
 
 import { render, screen } from '@testing-library/react';
-import { ThemaTitles } from '../../../universal/config/thema';
+import { ThemaTitles } from '../../config/thema';
 
 describe('<ThemaIcon/>', () => {
   it('Renders icon with label', () => {
